requirements.txt
streamlit>=1.28.0
supabase==2.3.0
websockets>=11.0.0
python-dotenv==1.1.1
pillow==10.4.0
tensorflow==2.14.0
numpy==1.25.2